// src/app/pages/artist/data-access/api.enum.ts
export const enum EndPointUrl {
    ARTISTS_CATEGORIES_ALL = "artists/categories/all",
    ARTISTS_CATEGORIES_CREATE = "artists/categories/create",
    ARTISTS_CATEGORIES_UPDATE = "artists/categories/update",
    ARTISTS_ALL = "artists/all",
    ARTISTS_ADD = "artists/add",
    ARTISTS_UPDATE = "artists/update",
    ARTISTS_DELETE = "artists/delete",
    ARTISTS_ROLES_ALL = "artists/roles/all",
}